+++ Sources/Contour/Transport/RabbitMQ/Internal/RabbitBusDefaults.cs
@@ -1,18 +1,15 @@
-﻿namespace Contour.Transport.RabbitMQ.Internal
+﻿using System;
+using Contour.Receiving;
+using Contour.Sending;
+using Contour.Transport.RabbitMQ.Topology;
+
+namespace Contour.Transport.RabbitMQ.Internal
 {
-    using System;
-
-    using Contour.Receiving;
-    using Contour.Sending;
-    using Contour.Transport.RabbitMQ.Topology;
-
     /// <summary>
     /// The rabbit bus defaults.
     /// </summary>
     internal static class RabbitBusDefaults
     {
-        #region Static Fields
-
         /// <summary>
         /// The route resolver builder.
         /// </summary>
@@ -23,9 +20,7 @@
         /// </summary>
         public static Func<ISubscriptionEndpointBuilder, ISubscriptionEndpoint> SubscriptionEndpointBuilder = SubscriptionEndpointBuilderImpl;
 
-        #endregion
-
-        #region Methods
+        public static IProducerSelectorBuilder ProducerSelectorBuilder = new DefaultProducerSelectorBuilder();
 
         /// <summary>
         /// The route resolver builder impl.
@@ -73,7 +68,5 @@
 
             return builder.ListenTo(queue, exchange);
         }
-
-        #endregion
     }
 }