+++ Sources/Contour/Transport/RabbitMQ/Internal/RabbitSender.cs
@@ -1,10 +1,15 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
-
 using Common.Logging;
-
+using Contour.Configuration;
 using Contour.Filters;
+using Contour.Helpers.CodeContracts;
 using Contour.Sending;
+using Contour.Transport.RabbitMQ.Topology;
 
 namespace Contour.Transport.RabbitMQ.Internal
 {
@@ -13,40 +18,36 @@ namespace Contour.Transport.RabbitMQ.Internal
     /// </summary>
     internal class RabbitSender : AbstractSender
     {
-        /// <summary>
-        /// Журнал выполнения отправителя.
-        /// </summary>
-        private static readonly ILog Logger = LogManager.GetCurrentClassLogger();
-
-        /// <summary>
-        /// Реестр поставщиков сообщений.
-        /// </summary>
-        private readonly ProducerRegistry producerRegistry;
-
-        /// <summary>
-        /// Поставщик сообщений.
-        /// </summary>
-        private Producer producer;
+        private readonly ILog logger;
+        private readonly RabbitBus bus;
+        private readonly IConnectionPool<IRabbitConnection> connectionPool;
+        private readonly ConcurrentBag<Producer> producers = new ConcurrentBag<Producer>();
+        private readonly RabbitSenderOptions senderOptions;
+        private IProducerSelector producerSelector;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="RabbitSender"/> class. 
         /// </summary>
-        /// <param name="endpoint">
-        /// Конечная точка, для которой создается отправитель.
+        /// <param name="bus">
+        /// A reference to the bus containing the sender
         /// </param>
         /// <param name="configuration">
         /// Конфигурация отправителя сообщений.
         /// </param>
-        /// <param name="producerRegistry">
-        /// Реестр поставщиков сообщений.
+        /// <param name="connectionPool">
+        /// A bus connection pool
         /// </param>
         /// <param name="filters">
         /// Фильтры сообщений.
         /// </param>
-        public RabbitSender(IEndpoint endpoint, ISenderConfiguration configuration, ProducerRegistry producerRegistry, IEnumerable<IMessageExchangeFilter> filters)
-            : base(endpoint, configuration, filters)
+        public RabbitSender(RabbitBus bus, ISenderConfiguration configuration, IConnectionPool<IRabbitConnection> connectionPool, IEnumerable<IMessageExchangeFilter> filters)
+            : base(bus.Endpoint, configuration, filters)
         {
-            this.producerRegistry = producerRegistry;
+            this.bus = bus;
+            this.connectionPool = connectionPool;
+            this.senderOptions = (RabbitSenderOptions)this.Configuration.Options;
+
+            this.logger = LogManager.GetLogger($"{this.GetType().FullName}(Endpoint=\"{this.bus.Endpoint}\")");
         }
 
         /// <summary>
@@ -61,7 +62,7 @@ namespace Contour.Transport.RabbitMQ.Internal
         {
             get
             {
-                return !producer.HasFailed;
+                return this.producers.Any(p => !p.HasFailed);
             }
         }
 
@@ -70,7 +71,7 @@ namespace Contour.Transport.RabbitMQ.Internal
         /// </summary>
         public override void Dispose()
         {
-            Logger.Trace(m => m("Disposing sender of [{0}].", this.Configuration.Label));
+            this.logger.Trace(m => m("Disposing sender of [{0}]", this.Configuration.Label));
             this.Stop();
         }
 
@@ -79,16 +80,15 @@ namespace Contour.Transport.RabbitMQ.Internal
         /// </summary>
         public override void Start()
         {
-            Logger.Trace(m => m("Starting sender of [{0}].", this.Configuration.Label));
-
             if (this.IsStarted)
             {
                 return;
             }
 
-            this.IsStarted = true;
+            this.logger.Trace(m => m("Starting sender of [{0}]", this.Configuration.Label));
 
-            this.EnsureProducerIsReady();
+            this.StartProducers();
+            this.IsStarted = true;
         }
 
         /// <summary>
@@ -96,10 +96,15 @@ namespace Contour.Transport.RabbitMQ.Internal
         /// </summary>
         public override void Stop()
         {
-            Logger.Trace(m => m("Stopping sender of [{0}].", this.Configuration.Label));
-            this.IsStarted = false;
+            if (!this.IsStarted)
+            {
+                return;
+            }
 
-            this.UnbindProducer();
+            this.logger.Trace(m => m("Stopping sender of [{0}]", this.Configuration.Label));
+
+            this.StopProducers();
+            this.IsStarted = false;
         }
 
         /// <summary>
@@ -109,60 +114,160 @@ namespace Contour.Transport.RabbitMQ.Internal
         /// <returns>Задача ожидания отправки сообщения.</returns>
         protected override Task<MessageExchange> InternalSend(MessageExchange exchange)
         {
-            this.EnsureProducerIsReady();
+            var producer = this.producerSelector.Next<Producer>();
 
             if (exchange.IsRequest)
             {
-                return this.producer.Request(exchange.Out, exchange.ExpectedResponseType)
+                return producer.Request(exchange.Out, exchange.ExpectedResponseType)
                     .ContinueWith(
                         t =>
-                            {
-                                if (t.IsFaulted)
-                                {
-                                    exchange.Exception = t.Exception;
-                                }
-                                else
-                                {
-                                    exchange.In = t.Result;
-                                }
-
-                                return exchange;
-                            });
-            }
-
-            return this.producer.Publish(exchange.Out)
-                .ContinueWith(
-                    t =>
                         {
                             if (t.IsFaulted)
                             {
                                 exchange.Exception = t.Exception;
                             }
+                            else
+                            {
+                                exchange.In = t.Result;
+                            }
 
                             return exchange;
                         });
+            }
+
+            return producer.Publish(exchange.Out)
+                .ContinueWith(
+                    t =>
+                    {
+                        if (t.IsFaulted)
+                        {
+                            exchange.Exception = t.Exception;
+                        }
+
+                        return exchange;
+                    });
         }
 
         /// <summary>
-        /// Гарантирует, что поставщик сообщений запущен.
+        /// Starts the producers
         /// </summary>
-        private void EnsureProducerIsReady()
+        private void StartProducers()
         {
-            if (this.producer == null)
+            this.logger.Trace(m => m("Starting producers for sender of [{0}]", this.Configuration.Label));
+            this.Configure();
+
+            foreach (var producer in this.producers)
             {
-                Logger.Trace(m => m("Resolving producer for sender of [{0}].", this.Configuration.Label));
-                this.producer = this.producerRegistry.ResolveFor(this.Configuration);
-                this.producer.Start();
+                producer.Start();
+                this.logger.Trace($"Producer of [{this.Configuration.Label}] started successfully");
+            }
+        }
+
+        private void Configure()
+        {
+            this.BuildProducers();
+            var builder = this.senderOptions.GetProducerSelectorBuilder();
+            this.producerSelector = builder.Build(this.producers.ToList());
+        }
+
+        /// <summary>
+        /// Stops the producers
+        /// </summary>
+        private void StopProducers()
+        {
+            this.logger.Trace(m => m("Stopping producers for sender of [{0}]", this.Configuration.Label));
+
+            Producer producer;
+            while (!this.producers.IsEmpty && this.producers.TryTake(out producer))
+            {
+                try
+                {
+                    producer.Stop();
+                    producer.Dispose();
+                    this.logger.Trace("Producer stopped successfully");
+                }
+                catch (Exception ex)
+                {
+                    this.logger.Error(
+                        $"Failed to stop producer [{producer}] in sender of [{this.Configuration.Label}] due to {ex.Message}");
+                }
             }
         }
 
         /// <summary>
-        /// Отписывается от поставщика сообщений.
+        /// Builds a set of producers constructing one producer for each URL in the connection string
         /// </summary>
-        private void UnbindProducer()
+        private void BuildProducers()
         {
-            Logger.Trace(m => m("Unbinding producer from sender of [{0}].", this.Configuration.Label));
-            this.producer = null;
+            var reuseConnectionProperty = this.senderOptions.GetReuseConnection();
+            var reuseConnection = reuseConnectionProperty.HasValue && reuseConnectionProperty.Value;
+
+            this.logger.Trace(
+                $"Building producers of [{this.Configuration.Label}]:\r\n\t{string.Join("\r\n\t", this.senderOptions.RabbitConnectionString.Select(url => $"Producer({this.Configuration.Label}): URL\t=>\t{url}"))}");
+
+            foreach (var url in this.senderOptions.RabbitConnectionString)
+            {
+                var source = new CancellationTokenSource();
+                var connection = this.connectionPool.Get(url, reuseConnection, source.Token);
+                this.logger.Trace($"Using connection [{connection.Id}] at URL=[{url}] to resolve a producer");
+
+                var topologyBuilder = new TopologyBuilder(connection.OpenChannel());
+                var builder = new RouteResolverBuilder(this.bus.Endpoint, topologyBuilder, this.Configuration);
+                var routeResolverBuilder = this.Configuration.Options.GetRouteResolverBuilder();
+
+                Assumes.True(
+                    routeResolverBuilder.HasValue,
+                    "RouteResolverBuilder must be set for [{0}]",
+                    this.Configuration.Label);
+
+                var routeResolver = routeResolverBuilder.Value(builder);
+
+                var producer = new Producer(
+                    this.bus.Endpoint,
+                    connection,
+                    this.Configuration.Label,
+                    routeResolver,
+                    this.Configuration.Options.IsConfirmationRequired());
+
+                producer.Failed += p =>
+                {
+                    // A failed producer will not be removed from the collection cause this failure should be treated as transient
+                    this.logger.Error($"Producer [{producer}] has failed");
+                };
+
+                if (this.Configuration.RequiresCallback)
+                {
+                    var callbackConfiguration = this.Configuration.CallbackConfiguration;
+
+                    this.logger.Trace(
+                        $"A sender of [{this.Configuration.Label}] requires a callback configuration; registering a receiver of [{callbackConfiguration.Label}] with connection string [{callbackConfiguration.Options.GetConnectionString()}]");
+
+                    var receiver = this.bus.RegisterReceiver(this.Configuration.CallbackConfiguration);
+
+                    this.logger.Trace(
+                        $"A new callback receiver of [{callbackConfiguration.Label}] with connection string [{callbackConfiguration.Options.GetConnectionString()}] has been successfully registered, getting one of its listeners with URL=[{producer.BrokerUrl}]...");
+
+                    var listener = receiver.GetListener(l => l.BrokerUrl == producer.BrokerUrl);
+
+                    if (listener == null)
+                    {
+                        throw new BusConfigurationException(
+                            $"Unable to find a suitable listener for receiver {receiver}");
+                    }
+
+                    this.logger.Trace(
+                        $"A listener at URL=[{listener.BrokerUrl}] belonging to callback receiver of [{callbackConfiguration.Label}] acquired");
+
+                    producer.UseCallbackListener(listener);
+
+                    this.logger.Trace(
+                        $"A producer of [{producer.Label}] at URL=[{producer.BrokerUrl}] has registered a callback listener successfully");
+                }
+
+                this.producers.Add(producer);
+                this.logger.Trace(
+                    $"A producer of [{producer.Label}] at URL=[{producer.BrokerUrl}] has been added to the sender");
+            }
         }
     }
 }